+++ module-variations-ui/_run.php
@@ -8,8 +8,8 @@ add_action( 'admin_init', '_run_admin_h_variations_ui' );
 /////
 
 function _run_admin_h_variations_ui() {
-  require_once HOO_PATH . '/module-variations-ui/variations-ui.php';
-  require_once HOO_PATH . '/module-variations-ui/template.php';
-  new H_VariationsUI();
-  new H_VariationsUI_Template();
+  require_once HOO_PATH . '/module-variations-ui/var-ui.php';
+  require_once HOO_PATH . '/module-variations-ui/var-template.php';
+  new \h\Variations_UI();
+  new \h\Variations_Template();
 }