+++ module-wholesale/wholesale-frontend.php
@@ -1,0 +1,5 @@
+<?php namespace h;
+
+class Wholesale_Frontend {
+  
+}